% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeperiod_functions.R
\name{create_intervalDT}
\alias{create_intervalDT}
\title{Convert integer sets to intervals}
\usage{
create_intervalDT(x, value_var, group_vars = NULL)
}
\arguments{
\item{x}{A data.table}

\item{value_var}{A length-1 character vector denoting the column of x which is the integer variable to be grouped into intervals}

\item{group_vars}{NULL or a character vector corresponding to column names of x. intervals will be grouped within categories defined by these columns}
}
\value{
a data.table with columns named "start" and "end" as well as group_vars if present.
The value variable must be unique within groups.
}
\description{
Given a set of discrete (integer) values, arrange adjacent values into inclusive intervals
Test whether a data.table contains intervals which overlap with other intervals in diferent rows, possibly within groups
}
\examples{
x <- data.table(value=c(1L,2L,3L,4L,10L,11L,12L,15L,1L,5L,6L),id=c(1L,1L,1L,1L,1L,1L,1L,1L,2L,2L,2L))
create_intervalDT(x,"value","id")
}
