% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeperiod_functions.R
\name{CJ.dt}
\alias{CJ.dt}
\title{a function to grid expand an arbitrary number of data.tables}
\usage{
CJ.dt(..., groups = NULL)
}
\arguments{
\item{...}{data.tables}

\item{groups}{a character vector corresponding to
column names of grouping vars in all of the data.tables}
}
\description{
similar to CJ
}
